+++ src/main/java/mod/emt/thaumictweaker/mixins/entities/EntitySpecialItemMixin.java
@@ -3,7 +3,6 @@ package mod.emt.thaumictweaker.mixins.entities;
 import net.minecraft.entity.item.EntityItem;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;